fix: fix redirect netlify rules
