+++ src/components/Map.jsx
@@ -145,7 +145,7 @@ const Map = () => {
             {searching ? 
                 <Spinner className='absolute top-0 right-2 h-full' aria-label="Default status example" /> 
                 : 
-                <div onClick={chooseCurrentLocation} className={`bg-transparent absolute top-2 bottom-2 right-2 w-6 bg-contain bg-center bg-no-repeat bg-[url("/location.svg")] hover:cursor-pointer  ${currentLocationChoosen ? 'bg-teal-200' : null}`}></div>
+                <div onClick={chooseCurrentLocation} className={`bg-transparent absolute top-2 bottom-2 right-2 w-6 bg-contain bg-center bg-no-repeat bg-[url("/assets/location.svg")] hover:cursor-pointer  ${currentLocationChoosen ? 'bg-teal-200' : null}`}></div>
             }
             <div className='bg-white w-full absolute top-[30px] shadow-xl rounded-b-md [&>*:nth-child(even)]:border-y-[1px] [&>*:nth-child(even)]:border-y-gray-300'>
                 {
