+++ src/components/TagSelector.css
@@ -1,0 +1,36 @@
+.tag_div{
+  display: flex;
+  flex-direction: column;
+}
+.tags{
+  display: flex;
+  flex-direction: row;
+  gap: 1rem;
+  flex-wrap: wrap;
+  padding: 1rem 0 2rem 0;
+}
+.tag_header{
+  padding-left: 0;
+  font-size: .9rem;
+}
+.each_tag{
+  padding: .5rem 1rem;
+  border-radius: 50px;
+  font-size: .8rem;
+  background: #3d5a802d;
+  width: fit-content;
+  cursor: pointer;
+  transition: .05s;
+}
+.each_tag:hover{
+  background: #3d5a80;
+  color: #f8f9fa;
+}
+.pick{
+  background: #133c55!important;
+  color: #f8f9fa!important;
+}
+.pick:hover{
+  background: #3d5a80!important;
+  color: #f8f9fa!important;
+}
